\docType{data}
\name{Codons}
\alias{Codons}
\title{Table of codons and encoded amino acids}
\format{A data frame with columns:
\describe{
 \item{ndiff}{Indicator, how many codons encode the same amino acid}
 \item{AA}{Amino acid three-lettre code}
 \item{name}{Amino acid full name}
 \item{seq}{Codon sequence}
}}
\usage{
Codons
}
\description{
Table of codons and encoded amino acids
}
\examples{
 Codons
}
\keyword{datasets}


