\name{createFilialSequencePopulation}
\alias{createFilialSequencePopulation}
\title{Generate new sequences by recombination}
\usage{
createFilialSequencePopulation(sequenceVector, generateNrecombinedSequences)
}
\arguments{
\item{sequenceVector}{Character vector of nucleotide sequences which will be used to create new sequences through recombination.}
\item{generateNrecombinedSequences}{Numeric value setting number of recombined sequences which will be generated}
}
\description{
This function generates new sequences from set of parental sequences through recombination.
}
\value{
Character vector of nucleotide sequences, generated by recombination from the entered \code{sequenceVector}, holding as much filial sequences as stated in \code{generateNrecombinedSequences}. Modes of recombination are cross-over, insertion and random.
}
\examples{
createFilialSequencePopulation(c('AAABBBCCCDDDEEEFFF','GGGHHHIIIJJJKKKLLL'), 3)

}
