\docType{data}
\name{hbg}
\alias{hbg}
\title{Donor sequences and their HBS}
\format{A data frame with columns:
\describe{
 \item{seq}{11nt long donor sequence}
 \item{hbs}{HBS of the donor sequence}
 \item{special_seq}{Shorter version of the donor sequence}
}}
\usage{
hbg
}
\description{
Donor sequences and their HBS
}
\examples{
 hbg
}
\keyword{datasets}


