\docType{data}
\name{hex}
\alias{hex}
\title{Hexamers and Z scores}
\format{A data frame with columns:
\describe{
 \item{seq}{Sequence of the hexamer.}
 \item{value}{ZEI-score of the hexamer from HEXplorer.}
 \item{first}{First codon within the hexamer.}
 \item{second}{Second codon within the hexamer.}
 \item{first_AA}{First encoded amino acid within the hexamer (three lettre code).}
 \item{second_AA}{Second encoded amino acid within the hexamer (three lettre code).}
  \item{AA}{Both encoded amino acid within the hexamer}
}}
\usage{
hex
}
\description{
Hexamers and Z scores
}
\examples{
 hex
}
\keyword{datasets}
